% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addSS}
\alias{addSS}
\title{Add necessary js components for shinyStorage}
\usage{
addSS()
}
\description{
\code{addSS}
A utility function to add the simpleStorage javascript library to a shiny app.
}
\examples{
\dontrun{
shinyUI({
fluidPage(
  addSS()
)
})
}
}

