% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{shinyStore-class}
\alias{shinyStore}
\alias{shinyStore-class}
\title{CLASS shinyStore}
\description{
class to handle local storage
}
\details{
This class handles local storage
}
\section{Fields}{

\describe{
\item{\code{session}}{An environment (\code{\link{session}}) associated with a shiny server function.}
}}

